+++ alby/models.go
@@ -17,7 +17,7 @@ type AlbyOAuthService interface {
 	GetBalance(ctx context.Context) (*AlbyBalance, error)
 	GetMe(ctx context.Context) (*AlbyMe, error)
 	SendPayment(ctx context.Context, invoice string) error
-	GetTopupUrl(ctx context.Context, amount int64, address string, currency string) ([]AlbyTopup, error)
+	GetTopupUrl(ctx context.Context, amount int64, address string, currency string) ([]AlbyTopupProvider, error)
 }
 
 type AlbyBalanceResponse struct {
@@ -39,7 +39,7 @@ type AlbyMe struct {
 	SharedNode       bool   `json:"shared_node"`
 }
 
-type AlbyTopup struct {
+type AlbyTopupProvider struct {
 	Name string `json:"name"`
 	Url  string `json:"url"`
 }
