+++ frontend/src/moonpay.ts
@@ -1,0 +1,142 @@
+export const MOONPAY_SUPPORTED_CURRENCIES = [
+  {
+    value: "usd",
+    label: "USD - US Dollar",
+  },
+  {
+    value: "ars",
+    label: "ARS - Argentine Peso",
+  },
+  {
+    value: "aud",
+    label: "AUD - Australian Dollar",
+  },
+  {
+    value: "bgn",
+    label: "BGN - Bulgarian Lev",
+  },
+  {
+    value: "brl",
+    label: "BRL - Brazilian Real",
+  },
+  {
+    value: "cad",
+    label: "CAD - Canadian Dollar",
+  },
+  {
+    value: "chf",
+    label: "CHF - Swiss Franc",
+  },
+  {
+    value: "cop",
+    label: "COP - Colombian Peso",
+  },
+  {
+    value: "czk",
+    label: "CZK - Czech Koruna",
+  },
+  {
+    value: "dkk",
+    label: "DKK - Danish Krone",
+  },
+  {
+    value: "dop",
+    label: "DOP - Dominican Peso",
+  },
+  {
+    value: "egp",
+    label: "EGP - Egyptian Pound",
+  },
+  {
+    value: "eur",
+    label: "EUR - Euro",
+  },
+  {
+    value: "gbp",
+    label: "GBP - Pound Sterling",
+  },
+  {
+    value: "hkd",
+    label: "HKD - Hong Kong Dollar",
+  },
+  {
+    value: "idr",
+    label: "IDR - Indonesian Rupiah",
+  },
+  {
+    value: "ils",
+    label: "ILS - Israeli New Shekel",
+  },
+  {
+    value: "jod",
+    label: "JOD - Jordanian Dollar",
+  },
+  {
+    value: "kes",
+    label: "KES - Kenyan Shilling",
+  },
+  {
+    value: "kwd",
+    label: "KWD - Kuwaiti Dinar",
+  },
+  {
+    value: "lkr",
+    label: "LKR - Sri Lankan Rupee",
+  },
+  {
+    value: "mxn",
+    label: "MXN - Mexican Peso",
+  },
+  {
+    value: "ngn",
+    label: "NGN - Nigerian Naira",
+  },
+  {
+    value: "nok",
+    label: "NOK - Norwegian Krone",
+  },
+  {
+    value: "nzd",
+    label: "NZD - New Zealand Dollar",
+  },
+  {
+    value: "omr",
+    label: "OMR - Omani Rial",
+  },
+  {
+    value: "pen",
+    label: "PEN - Peruvian Sol",
+  },
+  {
+    value: "pln",
+    label: "PLN - Polish Złoty",
+  },
+  {
+    value: "ron",
+    label: "RON - Romanian Leu",
+  },
+  {
+    value: "sek",
+    label: "SEK - Swedish Krona",
+  },
+  {
+    value: "thb",
+    label: "THB - Thai Baht",
+  },
+  {
+    value: "try",
+    label: "TRY - Turkish Lira",
+  },
+  {
+    value: "twd",
+    label: "TWD - Taiwan Dollar",
+  },
+  {
+    value: "vnd",
+    label: "VND - Vietnamese Dong",
+  },
+  {
+    value: "zar",
+    label: "ZAR - South African Rand",
+  },
+];
